选修进阶/媒体宝项目/code/mtb-master/requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==3.3
kombu==5.2.4
mysqlclient==2.1.0
openpyxl==3.0.9
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
pycryptodome==3.14.1
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
redis==4.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
